refactor: reorganize package structure and update imports for data models and entities
