+++ wallet/src/main/kotlin/net/thechance/wallet/api/dto/transaction/InitiateTransactionRequest.kt
@@ -1,8 +1,8 @@
 package net.thechance.wallet.api.dto.transaction
 
-import net.thechance.wallet.entity.InitiateTransactionParams
 import net.thechance.wallet.entity.Transaction
-import java.util.UUID
+import net.thechance.wallet.service.model.input.InitiateTransactionParams
+import java.util.*
 
 data class InitiateTransactionRequest(
     val receiverId: UUID,
