+++ wallet/src/main/kotlin/net/thechance/wallet/api/dto/transaction/TransactionResponse.kt
@@ -1,7 +1,7 @@
 package net.thechance.wallet.api.dto.transaction
 
 import net.thechance.wallet.entity.Transaction
-import net.thechance.wallet.service.helper.UserTransactionType
+import net.thechance.wallet.service.model.input.UserTransactionType
 import java.time.LocalDateTime
 import java.util.*
