+++ wallet/src/main/kotlin/net/thechance/wallet/service/model/input/TransactionFilterParams.kt
@@ -1,4 +1,4 @@
-package net.thechance.wallet.service.helper
+package net.thechance.wallet.service.model.input
 
 import net.thechance.wallet.entity.Transaction
 import java.time.LocalDate
