+++ wallet/src/main/kotlin/net/thechance/wallet/service/model/input/UserTransactionType.kt
@@ -1,4 +1,4 @@
-package net.thechance.wallet.service.helper
+package net.thechance.wallet.service.model.input
 
 enum class UserTransactionType {
     SENT,
