+++ wallet/src/main/kotlin/net/thechance/wallet/service/model/output/ReceiverDetails.kt
@@ -1,6 +1,7 @@
-package net.thechance.wallet.entity
+package net.thechance.wallet.service.model.output
 
-import net.thechance.wallet.entity.user.WalletUser
+import net.thechance.wallet.entity.Transaction
+import net.thechance.wallet.entity.WalletUser
 
 data class ReceiverDetails(
     val name: String,
@@ -9,7 +10,7 @@ data class ReceiverDetails(
 
 fun WalletUser.toReceiverDetails(type: Transaction.Type): ReceiverDetails {
     val isOnlinePurchase = (type == Transaction.Type.ONLINE_PURCHASE)
-    val receiverName = dukan?.name?.takeIf { isOnlinePurchase } ?: "$firstName $lastName"
+    val receiverName = dukan?.name?.takeIf { isOnlinePurchase } ?: userName
     val receiverImageUrl = if (isOnlinePurchase) dukan?.imageUrl else imageUrl
     return ReceiverDetails(
         name = receiverName,
