+++ wallet/src/main/kotlin/net/thechance/wallet/service/model/output/StatementData.kt
@@ -1,5 +1,6 @@
-package net.thechance.wallet.service.helper
+package net.thechance.wallet.service.model.output
 
+import net.thechance.wallet.service.model.input.UserTransactionType
 import java.time.LocalDateTime
 import java.util.*
